Added the menu function
