+++ galaga/StateMachine.cs
@@ -7,10 +7,8 @@ namespace galaga.GalagaStates {
         public StateMachine() {
             GalagaBus.GetBus().Subscribe(GameEventType.GameStateEvent, this);
             GalagaBus.GetBus().Subscribe(GameEventType.InputEvent, this);
-            Console.WriteLine("State 1");
             ActiveState = MainMenu.GetInstance();
             //ActiveState.InitializeGameState();
-            Console.WriteLine("State 2");
         }
         private void SwitchState(GameStateType stateType) {
             switch (stateType) {
